requirement.txt
myqueue==24.1.0
ase
clease==1.1.0
tqdm
toml
matplotlib
numpy
pandas
chgnet==0.4.0
mace-torch==0.3.13